requirements.txt
# Core dependencies
anthropic==0.7.2
pyyaml>=6.0.1
python-dateutil>=2.8.2

# Logging and error handling
rich>=10.0.0

# Type hints support
typing-extensions>=4.8.0

# Optional dependencies for enhanced functionality
matplotlib>=3.7.1  # For visualization if needed
pandas>=2.0.0     # For data handling
numpy>=1.24.0     # For numerical operations

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0

# Development dependencies
black>=23.7.0     # Code formatting
flake8>=6.1.0     # Code linting
mypy>=1.5.1       # Type checking
isort>=5.12.0     # Import sorting

# Documentation
sphinx>=7.1.2
sphinx-rtd-theme>=1.3.0
